e2e: extract home-screen helpers to fixtures/home.ts
Tasks 12 and 14 will reuse startMockingDefaultCity; pulling it out of
the test body keeps the LazyColumn-virtualisation and two-step start
quirks in one place.
