+++ e2e/tests/02-preset-mocking.test.ts
@@ -1,4 +1,3 @@
-import type { Device } from 'mobilewright';
 import { test, expect } from '../fixtures/device';
 import { expect as playwrightExpect } from '@playwright/test';
 import { Strings } from '../fixtures/selectors';
@@ -7,43 +6,7 @@ import {
   isForegroundServiceRunning,
   lastMockedLocation,
 } from '../fixtures/adb';
-
-// The LocationRow composable renders only the city portion of the name
-// (text = location.name.substringBefore(',')) so we match on just the city.
-const DEFAULT_CITY_LABEL = DEFAULT_CITY.name.split(',')[0].trim();
-
-/**
- * Select the default city and start mocking from any initial state.
- *
- * The hero card (with Start / Stop broadcasting) is at the top of the
- * LazyColumn. After scrolling to find Stockholm we scroll back up to
- * ensure the hero card is re-composed, then tap accordingly.
- */
-async function startMockingDefaultCity(device: Device): Promise<void> {
-  const { screen } = device;
-
-  // Scroll to the top in case the hero card has been virtualised away.
-  await screen.swipe('down');
-  await screen.swipe('down');
-
-  // If mocking is already active (leftover from a previous test), stop it first.
-  const stopCta = screen.getByText(Strings.home.stopBroadcastingCta);
-  if (await stopCta.isVisible({ timeout: 1_000 }).catch(() => false)) {
-    await stopCta.tap();
-    // Wait for the UI to return to idle state.
-    await screen.getByText(Strings.home.startCta).isVisible({ timeout: 5_000 });
-  }
-
-  // Scroll to Stockholm (off-screen in a sorted LazyColumn) and select it.
-  const cityLocator = screen.getByText(DEFAULT_CITY_LABEL);
-  await cityLocator.scrollIntoViewIfNeeded();
-  await cityLocator.tap();
-
-  // Scroll back to top and tap Start.
-  await screen.swipe('down');
-  await screen.swipe('down');
-  await screen.getByText(Strings.home.startCta).tap();
-}
+import { startMockingDefaultCity } from '../fixtures/home';
 
 test.describe('preset mocking', () => {
   test('tap preset transitions home to mocking-active state', async ({ device }) => {
